Support simple hardware accelerated raytracing
This supports the minimal required for raytracing the most simple opaque
triangular geometry with zero optimizations. For example, not even
D3D12_RAYTRACING_ACCELERATION_STRUCTURE_BUILD_FLAG_PREFER_FAST_TRACE can be
toggled yet. Also, acceleration structures cannot be updated yet, because
Hummingbird's scenes are currently static.

My first pass on this had specific and more complex RHI objects for bottom level
and top level acceleration structures which was too big of a break from the
"just wrap the APIs" approach that I want to take going forward where possible.
Now, the user can get the size of acceleration structures, create resources
with the appropriate flags, use a graphics context to schedule builds on the
GPU, and finally create a TLAS view themselves.

Similarly, I plan to fix deferred uploads, pending deletions, and
synchronization implicit in the RHI which fail the "just wrap the API" check.
These types of abstractions should be tailored by the user and for the user.